+++ pom.xml
@@ -9,7 +9,7 @@
 		<relativePath/> <!-- lookup parent from repository -->
 	</parent>
 	<groupId>com.</groupId>
-	<artifactId>restCrud</artifactId>
+	<artifactId>SpringbootPOC</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
 	<name>RESTCrud</name>
 	<description>Demo project for Spring Boot</description>
@@ -60,7 +60,12 @@
 			<version>5.9.2</version>
 			<scope>test</scope>
 		</dependency>
-	</dependencies>
+        <dependency>
+            <groupId>org.springdoc</groupId>
+            <artifactId>springdoc-openapi-common</artifactId>
+            <version>1.6.9</version>
+        </dependency>
+    </dependencies>
 
 	<build>
 		<plugins>
