--- src/main/java/com/restCrud/RestCrudApplication.java
@@ -1,17 +1,0 @@
-package com.restCrud;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-@SpringBootApplication
-@EnableJpaAuditing
-@ComponentScan(basePackages = "com.restCrud.*")
-public class RestCrudApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(RestCrudApplication.class, args);
-	}
-
-}
