--- src/main/java/com/restCrud/dbo/User.java
@@ -1,30 +1,0 @@
-package com.restCrud.dbo;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-
-
-/**
- * @author mamta.t
- */
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-@Entity
-public class User {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int userId;
-
-    private String username;
-    private String firstName;
-    private String lastName;
-    private String email;
-    private String password;
-}
