--- src/main/java/com/restCrud/exception/ResourceNotFoundException.java
@@ -1,21 +1,0 @@
-package com.restCrud.exception;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-/**
- * @author mamta.t
- */
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class ResourceNotFoundException extends RuntimeException{
-
-    String resourceName;
-    String fieldName;
-    long fieldValue;
-
-}
