+++ src/main/java/com/springbootPOC/SpringbootPOCApplication.java
@@ -1,0 +1,17 @@
+package com.springbootPOC;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+
+@SpringBootApplication
+@EnableJpaAuditing
+@ComponentScan(basePackages = "com.springbootPOC.*")
+public class SpringbootPOCApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(SpringbootPOCApplication.class, args);
+	}
+
+}
