+++ src/main/java/com/springbootPOC/constants/PageConstants.java
@@ -1,0 +1,14 @@
+package com.springbootPOC.constants;
+
+/**
+ * @author mamta.t
+ */
+public class PageConstants {
+    public static final Long DEFAULT_PAGE_NUMBER = 0L;
+
+    public static final Long DEFAULT_PAGE_SIZE = 10L;
+
+
+
+
+}
