+++ src/main/java/com/springbootPOC/dbo/User.java
@@ -1,0 +1,28 @@
+package com.springbootPOC.dbo;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+/**
+ * @author mamta.t
+ */
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+@Entity
+public class User {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private int userId;
+
+    private String username;
+    private String firstName;
+    private String lastName;
+    private String email;
+    private String password;
+}
