--- src/test/java/com/restCrud/RestCrudApplicationTests.java
@@ -1,13 +1,0 @@
-package com.restCrud;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class RestCrudApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
