+++ src/test/java/com/springbootPOC/SpringbootPOCApplicationTests.java
@@ -1,0 +1,13 @@
+package com.springbootPOC;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class SpringbootPOCApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
